salnutrition_back_end/pyproject.toml
[tool.poetry]
name = "salnutrition-back-end"
version = "0.1.0"
description = ""
authors = ["aleiaia <alessioiocca@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
manage = "django.core.management:execute_from_command_line"
